libguard: Add support for guard record history
Changes:
-Currently in libguard the guard records were wiped out
from the file, but changing it now to maintain the record
history. Hence instead of deleting, making the recordID
as FFFFFFFF and not displaying it.
-Change in create guard function, for marking correct recordId
because of change in delete function as we are marking recordId
as FF..FF it will take maxId is that and will not allocate correct
recordId to newly created guard record.
-Allowing to create guard of resolved guard record
-Adding function invalidateAll to invalidate all the guard records
i.e. mark recordId as FFFFFFFF for all the guard records present.
-Changing clearAll function to erase the data in guard file.
-Adding UT for invalidateAll function.
-Changes in Guard tool to not display invalidated guard records.

Test:
-Created guard record deleted the same, the recordId was setting
to FFFFFFFF and not getting displaying in guard tool output.
-Clearall is marking all recordId as FF...FF
-Created a guard record and deleted the same, tried to guard
the same record.

Signed-off-by: Chirag Sharma <chirshar@in.ibm.com>