libguard: Adding function to set/reset the write flag
Change:
Adding new function to set/reset the write flag.
While creating guard record or re-writing a guard record
write flag will be set to 0x00 and once the write completes,
set the flag to 0x01.

Tested:
While writing the guard data the flag is set as 0x00
and once the write is finished its set back as 0x01.
Created guard records to check if data is coming as
expected or not.

Signed-off-by: Chirag Sharma <chirshar@in.ibm.com>