Fixing logic error
This file has quite a few cpp core guideline violations.  In this case,
it appears to be doing manual pointer arithmetic, without an end bound,
which is explicitly disallowed and ideally should be handled by rnage
checked.  This commit does not attempt to fix all the issues, but simply
fixes the compiler error such that the OpenBMC rebase can continue
without hitting this compiler error.

In this case, the check appears to be checking a pointer operation
against nullptr, which is somewhat nonsensical, as a pointer incremented
by an offset is nearly always going to be non nullptr.

What I suspect his code was attempting to do was to see if the end of
the string, passed by a null terminated character was equal to null, ie
'\0'.

This commit changes the logic to do as I suspect the author intended,
which will allow the build to succeed.

Tested: Code compiles in latest openbmc rebase.  No guard system
available to test functionality.

Signed-off-by: Ed Tanous <edtanous@google.com>