Validate core guard clearing
On trying to clear/invalidate/delete individual guard record, we need to
validate if clearing is allowed for that record. An individual manual/
system generated core record can be removed. But other target guard
records cannot be deleted if it is system generated.

A Force delete can also be performed using the flag provided in the API.
The guard CLI tool uses this to invalidate any type of record.

guard -i <guard record> force clears the guard record.

This commit addresses the logic to handle it.

Tested and works as expected.

Signed-off-by: Deepa Karthikeyan <deepakala.karthikeyan@ibm.com>